+++ converter/src/test/kotlin/no/entur/geocoder/converter/source/adresse/MatrikkelConverterTest.kt
@@ -203,7 +203,11 @@ class MatrikkelConverterTest {
 
         lines.forEach { line ->
             val nominatimPlace: NominatimPlace = objectMapper.readValue(line)
-            val importance = nominatimPlace.content.first().importance
+            val importance =
+                nominatimPlace.content
+                    .first()
+                    .importance
+                    .toDouble()
 
             assertTrue(importance > 0.0, "Importance should be positive")
             assertTrue(importance <= 1.0, "Importance should not exceed 1.0")
